main: Downgrade client version when server is to recent
When the server is too recent the Docker client cannot communicate
properly. In such case the client shall degrade its version and
reconnect with the server's version.

To achieve that, once we get a communication channel with Docker, an
Info() command is performed to see if the protocol match.

- If not, the client version is overrided and the communication is
restarted.
- If it does, the communication channel is valid and the program continue.

dockerAPIVersion & dockerExist functions are then deprecated.