release: Adding makefile to build versionned binary
When building the binary, the actual code was embedding a git hash as a
version.

This is particulary wrong since the code representing that hash wasn't
the one you were running since you could have local modifications.

This patch create a version string based on the latest tag + branch name
+ actual commit.

This value is computed at build time by passing some "ldflags -X"
arguments. The value are computed from the git cli command.

A makefile have been created to ease this building make.
A simple "make" call is enough to get an up to date binary.

A typical output is like :
    ceph-nano version v1.1.0 (master/94d9d45)
    ceph-nano container image version 3d2c81b