release: Building versionned binaries
When build cn with the makefile, the previous binary is always
overwritten.

This patch integrate the current tag & version in the filename and makes
a symlink to cn. It's easier to call 'cn' when developping.

When calling clean target, only the symlink is deleted.
Calling clean-all target does remove all the binaries

Note that the TAG is set by default to "devel" as a TAG points precisely
to a particular version of the code. Keeping it here could be confusing.

The release process will take care of putting the proper tag in place.