tests: Creating reportSuccess() function
The actual code is considering reporting a success or failure based on
the result of a single 'cn' command.

runCn is now saving the return code into the runCnStatus variable and
let the user make a call to reportSuccess() once the testing function is done.

The error handling is now implicit : if a command fails the error handling is automatically engaged.

This will allow more complex tests by running several cn commands before returning a success.