tests: Adding isS3ObjectExists
When performing put/del/mv commands, its not enough to get the result of
the s3 command. To get a a real functional test, it is necessary to make
a functional check if the action have been properly performed.

This patch does add isS3ObjectExists() to check if a given file exists
in a bucket.

This function requires runCn to be verbose for parsing its output.
The new runCnVerbose variable can be used to get runCn being verbose.