tests: Improving output
The actual code was printing the name of the running task only once it succeed or failed.
Some tasks could take a while and make a confusing output where nothing prints for a long time.

This patch improves two things :
- it does print the name of the running test at its start and wait for a success/failure
- prints the running time once it succeed

This means that every test should call the "start_test" function to
print the running task & get the initial timestamp.

A typical output looks like (note that test_purge is actually running):
    [...]
    test_restart         : SUCCESS : 35,23 seconds
    test_status          : SUCCESS : 0,27 seconds
    test_version         : SUCCESS : 0,03 seconds
    test_update          : SUCCESS : 3,13 seconds
    test_purge           :

Note that if a test call another test (nested tests), only the first
test will be reported.