start: rework selinux implementation
When running on Linux and doing bindmounts for Ceph we must have the
right Selinux labels when running on a dedicated directory. Otherwise
the osd bootstrap will fail.

Also there is no need for a validation earlier, the working directory
does not need any SeLinux label to store and put data in the container.

Even more, the initial implementation wasn't working at all so I fix
that as well, annoyingly cmd.Exec returns empty line so we have to loop
over the result to parse and do the condition on Selinuw == Enforcing.

Signed-off-by: Sébastien Han <seb@redhat.com>