s3: fix ls and la calls
Prior to this commit the output was incorrect and also the message "on
cluster $cluster" wasn't approprieate so I removed it.

Now 'ls' can receive a single argument only in order to list all the
buckets if empty where 'la' will continue to report all the objects in
all the buckets.

Signed-off-by: Sébastien Han <seb@redhat.com>