functional-tests: Removing extra reportSuccess
Commit 4da0bcb6c32e470e274e7c978234aceb854a931b introduced an extra
reportSuccess call in test_start()

As per the nasted implementation, every reportSuccess is tied to a
start_test. This allow reporting only the main test name reported when
it calls many sub tests.

This extra reportSuccess was breaking this rule and put all the subtests
to be reported as main test. That made the debugging more difficult as
it was impossible to get what "main" test was failing.

Signed-off-by: Erwan Velu <erwan@redhat.com>