functional-tests: Avoid hardcoded cluster name
Commit 4da0bcb6c32e470e274e7c978234aceb854a931b introduced the
multi-cluster support but was hardcoding the number of cluster to
creates and the default cluster name to use.

To avoid any typo, this patch put the cluster name as a variable and so
for the number of clusters to manage.

A new variable "current_cluster_name" is also introduced to let every
runCn command selecting the current cluster.

Signed-off-by: Erwan Velu <erwan@redhat.com>