functional-tests: Improving delta error message
Some tests computes a delta between the number of items to handle and
the one actually existing.

As per the nested tests, it was almost impossible to understand what
sub-test was generating the delta error as all delta error message where
similar.

A typical output was like :
    test_s3_mv_50x_after_copy           : ERROR  : delta is 0

This patch offer to add the actual function name in the error message to
help debugger understanding what sub-function failed.

The resulting output is now like :
    test_s3_mv_50x_after_copy           : ERROR  : test_s3_mv_custom: delta is 0

Signed-off-by: Erwan Velu <erwan@redhat.com>