functional-tests: Cleanup clusters when failing
If the tests fails, it's important to cleanup the created clusters
unless the next start will fail.

This patch removes the cleanup from travis as it doesn't really know
what were the created clusters.

This patch also add some code in the fatal() function to cleanup any
possible clusters in case of failing.

Signed-off-by: Erwan Velu <erwan@redhat.com>