Makefile: Avoid undefined version in binary
The reported version was permanently set to "undefined/undefined" while
the expected behavior was to report a mix of version, tag & branch.

The makefile is using the ldflags with a -X to override variables.
The format is like <pkg>.<varname>=value.

It appear that if we don't pass the main.go on the command line, the
package isn't considered making the variable not being overrided.

This patch adds the "main.go" from the root of the project on the build
command line to get the package being considered.