utils: Improving json loader
The actual implementation is trying to workaround an utf8 conversion by
removing arbitraty parts of the loaded string which represent a Json.

This commit is doing two things :
- first it removes control & escape codes from the utf8 string
  This code is commit from the rosettacode project
- For an unknown reason, there is a left over char before the Json
  This commit removes everything before the first openning bracket '{'

This commit isn't perfect as it doesn't provide the answer for those three questions :
- why using strings.trimleft() doesn't work on that filtered string.
  We implemented after() as a workaround
- Why can't load directly the Json under its utf8 format ?
- Why do we have some left-over chars before the Json after calling stripCtlAndExtFromUTF8() ?

At least, this commit is a little bit as it removes arbitrary values.

Signed-off-by: Erwan Velu <erwan@redhat.com>