release.sh: Don't release if tests fails
When releasing the product, it is mandatory that at least the
functional tests from the build arch are passing. If not, let's stop the
release process.

The getGoEnv function extracts the os & arch from golang to detect which
freshly built binary have to be used to run the functional tests.

Signed-off-by: Erwan Velu <erwan@redhat.com>