update-check: Don't fail if API rate limit is exceeded
When running many PRs on travis, we sometimes reach the API rate limit like in :
    API rate limit exceeded for w.x.y.z (But here's the good news: Authenticated requests get a higher rate limit. Check out the documentation for more details.)

This patch simply check if a message is present in the answer, if so,
let's print it and return.

update-check is not really critical so it's acceptable just reporting the github message instead of failing.

Signed-off-by: Erwan Velu <erwan@redhat.com>