utils: grepForSuccess show log since 30s
We must show logs that are 30 seconds old. 30 sec represents the average
time for the container to bootstrap. If you restart the container, the
logs are not flushed, so if we don't limit the amount of logs to grep we
will obviously grep the initial SUCCESS occurence (representing the
initial invocation/bootstrap of cn). A restart might not succeed and the
only way to have grepForSuccess working correctly is to grep through the
latest logs.

Signed-off-by: Sébastien Han <seb@redhat.com>