Packaging: Adding Gopkg.toml
The golang packaging requires having a Gopkg.toml.
This file was generated by "dep init".
By default, the docker/docker was set to 1.13 which is too old regarding the current cn code.
It got updated to v17.05.0-ce.

The makefile is updated accordingly to use 'dep ensure' instead of 'go get'.
This guarantee the versionning to be fullfiled.

Travis is also updated to have go-dep installed so the "dep" command is available.
Installing dep from github as go-dep is only available from bionic which is not yet avail on travis.
Let's also ensure we run of the latest ubuntu to get a recent go version.

Signed-off-by: Erwan Velu <erwan@redhat.com>