start: start the mon on localhost
We have been having issues where the container restarts and pick a
different address. This is a problem for the monitor has it will not be
able to listen on the register IP from the monmap anymore. Using
localhost (127.0.0.1) ensures the mon will always find the IP. Since we
don't expose the monitors this is not really an issue.

Also, the network has been adapted to 0.0.0.0/0 so osds will bind on any
of the interfaces they find.

Signed-off-by: Sébastien Han <seb@redhat.com>