Packaging: Simplifying versionning
Up to now at every build, the version, tag & branch should be given.
In a development context, that make sense but when considering building a released version, that's is meaningless.

In the code, those 3 fields were merged into a single string : tag + " (" + branch + "/" + version + ")

To simplify the packaging, this patch offer to pass a single string.
The makefile, in a development mode, will make an output like $(TARGET_BRANCH)-$(COMMIT) : cn-master-cea247c-dirty-linux-amd64
If the VERSION string is override a build time, the user can pass anything. This is what will be used during the release process or a downstream build.

As the commit-id is removed from the filename, this commit also removes the commit string manipulation from the travis script.

The README is adjusted temporarly to mimic the expected format so the travis script works fine in the next release.

Signed-off-by: Erwan Velu <erwan@redhat.com>