contrib/travis: Updating specfile when releasing
When a release is made, the version is bumped but the specfile should be updated accordingly.

The tune-spec script will update the current spec file by :
- update the version
- update the release
- add a new changelog entry in the proper format
The changelog is the same as per the Release notes but formated for the spec file.
Once its updated, the spec file is commited the same way its done for the README.

This commit introduce
- setup_git       function to preparing the git env is generic
- commit_and_push function so the commit process is shared between 'commit_changed_readme' and 'commit_specfile'

Signed-off-by: Erwan Velu <erwan@redhat.com>