cmd: Adding configuration file support
This patch is about adding the option to add configuration files for
defining some tuning.

The configuration file could be placed in /etc/ceph or in the home
directory and should be named "cn.toml".

The tuning parameters are considered in this order :
- if no configuration file is present and if the parameter is mandatory, ceph-nano uses its internal hardcoded value.
- if there is a configuration file:
  - if there is any internal harcoded value, that's the default value
  - if the parameter is re-defined in the [default] section, this value superseed the internal hardcoded value
  - if the parameter is re-defined in the [myclustername] section,  this value superseed any other one.

Let's consider the first implementation in this commit, the size of the container memory.

The current default value is 512MB and harcoded directly in container.Resources.memory field.

The new scheme is to define its default value in setDefaultConfig() which creates a void entry in the [default] section with this 512MB value.
So mimic a user adding an entry like the following in the configuration file :

    [default]
      MemorySize=512MB

When ceph-nano needs to pass the memory size option to docker, it uses a call to getMemorySize(containerName).
This function will extract with getValueFromConfig() the current value for this containerName.
If the container doesn't define a particular value, the [default] value will be used.

This commit also add a test framework to guarantee a stable behavior when patching this part of the code.

Signed-off-by: Erwan Velu <erwan@redhat.com>