config: Implement use_default
The automatic inheriting of [default] is implicit (use_default=true) but
could be disabled in some particular cases.

If use_default=false, then a specific container should not inherit from the default value of a given parameter but neither reuse all the other defined values.

This patch is about implementing how retrieving a parameter should be
done regarding the state of use_default.

Signed-off-by: Erwan Velu <erwan@redhat.com>