config: Reading ceph.conf parameters from config file
A container can define a [<containername>.ceph.conf] section to define a list of parameters to be passed to the ceph.conf configuration file.

This commit add a getStringMapFromConfig() function to read a set of parameters at once while taking care of the use_default status.

The getCephConf() func implement the call to getStringMapFromConfig() to retrieve the "ceph.conf".

This commit implements the associates tests accordingly.

Signed-off-by: Erwan Velu <erwan@redhat.com>