config: Report memory & cpu settings at start time
When starting a cluster, it's useful to know what is the memory/cpu tuning in place.
To get a human readable output, this patch rename getMemorySize() to getMemorySizeInBytes().

A new function getMemorySize() reports the value as its written in human format in the config file.

A typical output looks like :
    [user@host] cn cluster start test_nano_no_default
        Using /home/erwan/.cn/cn.toml as configuration file
        2018/12/07 12:50:43 Running cluster test_nano_no_default (1GB Memory / 2 CPU) ...

This commit updates the tests for getMemorySizeInBytes and create new
ones for the new version of getMemorySize()

Signed-off-by: Erwan Velu <erwan@redhat.com>