config: Adding flavor types to containers
When starting a container, the configuration file provides a set of default flavors to be used.

The new "-f" option allow the user to select one or "default" will be used.
"default" represent the builtin values of ceph-nano or the overrided ones in the [default] section of the configuration file.

A check is done at startup to ensure the requested flavor exists.

To be consistent with this vision, the containerName variables in the code are renamed to containerFlavor.

This commit also brings a potential default configuration file (cn.toml), to be adjusted over time, to offer a set of default flavors to be used.

Signed-off-by: Erwan Velu <erwan@redhat.com>