config: Adding [images] support
Without a configuration file, the "-i" option allow users to choose a particular container to be used.
This behavior is kept for compatibility.

When running with a configuration file, the -i option points to an entry into the configuration file.

    cn cluster start -i redhat

In this example, the configuration file have this redhat entry defined as in :
[images]
  [images.redhat]
    image_name="registry.access.redhat.com/rhceph/rhceph-3-rhel7"

The image_name variable is used as input parameter to docker to pull the right image.
This configuration option makes easier to remember/use some complex/long urls.

This option can be combined with the flavor (-f) option to select the proper size of container.
It is then possible to run :

    cn cluster start -i redhat -f huge
or
    cn cluster start -i redhat -f medium

Signed-off-by: Erwan Velu <erwan@redhat.com>