start: Removing privileged option from command line
The privileged option was exposed to the user via the cli but didn't really meant to be directly used.
When needed this value was set directly by the code itself.

This patch is about to remove the option from the cli while keeping an option to enforce it in the flavors.
This could let an option to reenable it quickly if some need it.
After some time, if that is no more used, this could be totally removed.

The default value remain set to false.

This commit also add getBoolFromConfig() to retrieve a bool from the configuration.
It also add the needed tests to guarantee stability over time.

Signed-off-by: Erwan Velu <erwan@redhat.com>