config: Add "--data" support in flavors
When defining a flavor, it's possible to define the default value of the "--data" cli option.
The flavor can define a "data" paramater to define it.
If the use pass the '--data' on the command line, this one will be prefered over the flavor defined one.

The internal function to read this value is called getUnderlyingStorage() to keep a good semantic.
The flavor parameter is kept to "data" to mimic the option name.

Signed-off-by: Erwan Velu <erwan@redhat.com>