utils: Unifying toBytes() conversions
getMemorySizeInBytes() is using units.ParseBase2Bytes() to convert human readable units into bytes.
The existing code was using a toBytes() function to perform the same operation.

This commit is about having a single function relaying on units.ParseBase2Bytes().
This way the code is more robust & shared between functions.

This functions is tested with getMemorySizeInBytes() checks.

Signed-off-by: Erwan Velu <erwan@redhat.com>