config: Reworking merging of default & other flavors
When an item is defined in [default] and if a flavor keep the
"use_default=true" parameter, the item should be present in the flavor.

This patch is adding a new_param=true in the testing cn file so it
should be found in [test_nano_default] flavor and not in the [test_nano_no_default] one.

This new test proven this feature to be broken.

This patch does :
- switch use_default to a real bool and no more a string

- add mergeFlavorsWithDefault() function
  called by ReadConfigFile() to make all items of [default] being populated in other flavors when use_default is set to true.

- add getDefaultParameters() to list all parameters from [default]
  The default parameters aren't reachable from a Get() call, so a workaround is needed to get them.

- remove all the default parameter extraction from get*FromConfig()
  This makes the code simplier & easier to update

Signed-off-by: Erwan Velu <erwan@redhat.com>