config: Adding more parallelism in tests
The tests were written to be run a single time but it appear that running some of them several time broke the suite.

One of the reason is that once the ReadConfigFile() is done, the context is changed and doesn't really represent a default status.

This commit is about :
- having a prepare target that tries to run the test 5 times
- split all the tests on the default config to be run first
- split all the tests on the config file after a single read of the config file
- ensure that if some internal variables are used to triggering some particular behaviors, variables are restored after the test

This patch made the tests being more robust, it is now possible to run them 5 times without any issue.

Signed-off-by: Erwan Velu <erwan@redhat.com>