list: Reporting cluster flavor
When listing clusters, it could be very useful to report the selected flavor at start time.
To do so, dockerInspect() must read the Labels from the running container.
The showNanoClusters then, have to report this value to the table displayed to the user.

This patch is then adding the "flavor" support argument to dockerInspect() but also refactor it to use the switch() statement.
Using switch() statement makes the code easier to read & to expand.

If not flavor is found, already running containers, the "none" value is printed.

A typical output looks like :

$ ./ cn cluster ls
+---------------+---------+------------------------------+----------------+--------------------------------+--------+
| NAME          | STATUS  | IMAGE                        | IMAGE RELEASE  | IMAGE CREATION TIME            | FLAVOR |
+---------------+---------+------------------------------+----------------+--------------------------------+--------+
| monpetitponey | running | docker.io/ceph/daemon:latest | master-7150c5c | 2018-08-31T12:40:14.676023593Z | medium |
+---------------+---------+------------------------------+----------------+--------------------------------+--------+

Signed-off-by: Erwan Velu <erwan@redhat.com>