contrib: Improving bash_completion generation
Cobra offer a way to generate the bash_completion script.

This commit implement a new option in ceph nano (completion) to generates this script.
Calling 'cn completion' will render the script file to stdout.

This commit call this mecanism from the travis script so every release will update this file.

The spec file is updated to feature this file in the package.

The completion file is renamed to cn_completion.sh to have a more explicit name when located in /etc/bash_completion.d/

Signed-off-by: Erwan Velu <erwan@redhat.com>