flavors: Improving 'flavors show' output
When reporting the details of a flavor, it could be better and easier to
import into a 3rd party tool to only print the content of the flavor.

The introduction message was removed.
The flavor name feature a 'name' field which is automatically feed with the flavor name.
A test is added accordingly to check this feature is working properly.

Signed-off-by: Erwan Velu <erwan@redhat.com>