Makefile: go test must be run in the local context
When running a make release, that imply cross-compiling for other arch/Oses.
But once GOOS & GOARCH is defined, the tests were exec'd in such context leading to this error :
    fork/exec /tmp/go-build035358901/github.com/ceph/cn/cmd/_test/cmd.test: exec format error

This patch is about unsetting GOOS & GOARCH during the go test run so the go context is the host one and not the cross-compiled one.

Signed-off-by: Erwan Velu <erwan@redhat.com>