cmd/utils: Fix pagination on image list
The current page count value is using the tag count divided by 10. But
we are using a page size value of 100.
That means we're iterate 10x more than we should do.
Also we should add an extra page when the remainder of the modulo
operation between the tag count and page size isn't zero.

With:
  tagCount = 1937
  pageSize = 100
Then:
  pageCount = 20

Because there's 19 pages with 100 results and 1 page with 37 results.

Adding the pageSize as a constant.

$ cn image ls -a
before : ~ 22 seconds
 after : ~  6 seconds

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>